splitting up css, deleting styles that are not used
