+++ demo/src/components/highlight/HighlightContainer.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import '../../css/HighlightContainer.css';
 
 /*******************************************************************************
   <HighlightContainer /> Component
