+++ demo/src/css/App.css
@@ -1,0 +1,168 @@
+/* Base Styles */
+
+body,
+html {
+  min-width: 48em;
+  background: #f9fafc;
+  font-size: 16px;
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
+
+* {
+  font-family: 'Source Sans Pro', sans-serif;
+  color: #232323;
+}
+
+#root,
+#root > div {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+section {
+  background: #fff;
+}
+
+code,
+code span,
+pre {
+  font-family: 'Roboto Mono', monospace !important;
+}
+
+code {
+  background: #f6f8fa;
+}
+
+pre > code {
+  font-size: .8125em;
+  line-height: 1.5em;
+  padding: 1.875em;
+  display: block;
+  color: #333;
+  overflow-x: auto;
+}
+
+ol,
+ul {
+  font-size: 16px;
+}
+
+li,
+p,
+td,
+th {
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  font-size: 1.125em;
+  line-height: 1.5em;
+  margin: 1.2em 0;
+}
+
+p code {
+  font-size: .78125em;
+  padding: .125em .375em;
+}
+
+pre {
+  margin: 2em 0;
+}
+
+h1,
+h2 {
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  font-weight: 300;
+}
+
+h4,
+h5,
+h6 {
+  font-size: 1em;
+  margin: 1.2em 0;
+}
+
+h2 {
+  font-size: 2em;
+  color: rgba(35, 35, 35, .75);
+}
+
+a {
+  text-decoration: none;
+  cursor: pointer;
+  color: #2085bc;
+  font-weight: 700;
+  transition: color .2s ease, opacity .2s ease;
+}
+
+a:hover {
+  color: #40affd;
+}
+
+a:active {
+  opacity: .66;
+  color: #2085bc;
+  -webkit-transition-duration: 0s;
+  transition-duration: 0s;
+}
+
+blockquote {
+  border-left: #f6f8fa solid .375em;
+  padding-left: 1.125em;
+}
+
+blockquote * {
+  color: rgba(35, 35, 35, .75);
+}
+
+img {
+  max-width: 100%;
+}
+
+hr {
+  display: block;
+  border: none;
+  height: .375em;
+  background: #f6f8fa;
+}
+
+blockquote,
+hr {
+  margin: 2.4em 0;
+}
+
+table {
+  width: 100%;
+  border-collapse: collapse;
+  border: none;
+}
+
+tr:nth-child(even) {
+  background: rgba(184, 198, 213, .13);
+}
+
+td,
+th {
+  text-align: left;
+  padding: 1.25em;
+}
+
+td sub,
+th sub {
+  vertical-align: baseline;
+  padding-left: .375em;
+  opacity: .5;
+  font-size: 75%;
+}
+
+th {
+  border-bottom: .375em solid rgba(184, 198, 213, .13);
+}
+
+.u-hidden {
+  display: none;
+}
