+++ demo/src/css/Button.css
@@ -1,0 +1,52 @@
+.btn {
+  text-decoration: none;
+  cursor: pointer;
+  text-transform: uppercase;
+  font-size: 1em;
+  margin: 0;
+  -moz-appearance: none;
+  -webkit-appearance: none;
+  border: none;
+  color: #fff!important;
+  display: block;
+  background: #2085bc;
+  padding: .9375em 3.625em;
+  -webkit-transition: background-color .2s ease, opacity .2s ease;
+  transition: background-color .2s ease, opacity .2s ease;
+}
+
+.btn.btn--blue {
+  background: #2085bc;
+}
+
+.btn:focus,
+.btn:hover {
+  background: #40affd;
+  outline: 0;
+}
+
+.btn:focus {
+  box-shadow: 0 0 1.25em rgba(50, 50, 150, .05);
+}
+
+.btn:active {
+  opacity: .66;
+  background: #2085bc;
+  -webkit-transition-duration: 0s;
+  transition-duration: 0s;
+}
+
+.btn.btn--icon-disclosure svg {
+  width: .4375em;
+  height: .75em;
+  fill: #fff;
+  opacity: .66;
+  margin: 0 -.75em 0 .75em;
+}
+
+.btn:disabled,
+.btn:disabled:active,
+.btn:disabled:hover {
+  cursor: default;
+  background: #d0dae3;
+}
