+++ demo/src/css/HighlightContainer.css
@@ -1,0 +1,15 @@
+.highlight-container {
+  line-height: 42px !important;
+  align-items: center;
+  display: flex;
+  flex-wrap: wrap;
+  white-space: pre;
+  cursor: default;
+}
+
+/* Bottom Label Layout */
+
+.highlight-container.highlight-container--bottom-labels {
+  padding: 10px 1.125em;
+  align-items: flex-start;
+}
