+++ demo/src/css/Pane.css
@@ -1,0 +1,172 @@
+.pane-container {
+  display: flex;
+  height: 100%;
+}
+
+.pane {
+  display: flex;
+  flex-direction: row;
+  position: relative;
+  flex: 2;
+  height: auto;
+  min-height: 100%;
+  min-height: 34.375em;
+}
+
+.pane__horizontal {
+  display: block;
+}
+
+.pane__left,
+.pane__right,
+.pane__top,
+.pane__bottom {
+  width: 100%;
+  height: 100%;
+  align-self: stretch;
+  min-width: 24em;
+  min-height: 34.375em;
+}
+
+.pane__top {
+  height: auto;
+  display: block;
+  min-height: 0;
+}
+
+.pane__bottom {
+  min-height: 0;
+  height: auto;
+  display: block;
+  position: relative;
+}
+
+.pane__bottom.model__output--empty .placeholder {
+  min-height: 250px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.pane__left {
+  height: auto;
+  min-height: 100%;
+}
+
+.pane__right {
+  width: 100%;
+  overflow: auto;
+  height: auto;
+  min-height: 100%;
+}
+
+.pane__right .model__content.model__content--srl-output {
+  display: inline-block;
+  margin: auto;
+}
+
+.pane__thumb {
+  height: auto;
+  min-height: 100%;
+  margin-left: -.625em;
+  position: absolute;
+  width: 1.25em;
+  top: 0;
+}
+
+.pane__thumb:after {
+  display: block;
+  position: absolute;
+  height: 100%;
+  top: 0;
+  content: "";
+  width: .25em;
+  background: #e1e5ea;
+  left: .5em;
+}
+
+.pane__bottom .pane__thumb {
+  min-height: 0;
+  height: 0;
+  position: relative;
+  margin-left: 0;
+  margin-top: 0;
+  width: 100%;
+}
+
+.pane__bottom .pane__thumb:after {
+  height: .25em;
+  top: 0;
+  width: 100%;
+  left: 0;
+}
+
+/* Placeholder */
+
+.placeholder {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  user-select: none;
+  -webkit-touch-callout: none;
+  cursor: default;
+}
+
+.placeholder .placeholder__content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+}
+
+.placeholder svg {
+  display: block;
+}
+
+.placeholder svg.placeholder__empty,
+.placeholder svg.placeholder__error {
+  width: 6em;
+  height: 3.625em;
+  fill: #e1e5ea;
+  margin-bottom: 2em;
+}
+
+.placeholder svg.placeholder__error {
+  width: 4.4375em;
+  height: 4em;
+}
+
+.placeholder p {
+  font-size: 1em;
+  margin: 0;
+  padding: 0;
+  color: #9aa8b2;
+}
+
+.placeholder svg.placeholder__working {
+  width: 3.4375em;
+  height: 3.4375em;
+  -webkit-animation: working 1s infinite linear;
+  animation: working 1s infinite linear;
+}
+
+@-webkit-keyframes working {
+  0% {
+    -webkit-transform: rotate(0deg);
+  }
+  100% {
+    -webkit-transform: rotate(360deg);
+  }
+}
+
+@keyframes working {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
