+++ demo/src/css/hierplane-overrides.css
@@ -1,0 +1,76 @@
+.hierplane__visualization-verbs {
+  background: #f9fafc;
+  padding: 1em 1.25em;
+  border-bottom: 1px solid #e5eaf0;
+  color: rgba(28, 47, 58, .5);
+}
+
+.hierplane__visualization-verbs a {
+  display: inline-block;
+  margin-right: 1em;
+}
+
+.hierplane__visualization-verbs a svg {
+  fill: rgba(28, 47, 58, .5);
+}
+
+.hierplane__visualization-verbs a:hover svg {
+  fill: #1c2f3a;
+}
+
+.hierplane__visualization-verbs__prev {
+  transform: rotateY(-180deg);
+}
+
+.hierplane__visualization-verbs__label {
+  user-select: none;
+}
+
+/* THESE ARE HACKS PUT HERE BY CODEVIKING. */
+
+/*
+  Center the controls in the toolbar.
+ */
+.hierplane__visualization .hierplane .parse-tree-toolbar .parse-tree-toolbar__item .parse-tree-toolbar__item__label {
+  top: 50%;
+  margin-top: -13px;
+}
+
+/*
+  Make the toolbar background transparent, as it looks weird when it matches the color of the
+  tab / verb-select bars.
+ */
+.hierplane__visualization .hierplane--theme-light .parse-tree-toolbar {
+  background: transparent;
+  border: none;
+}
+.hierplane__visualization .hierplane--theme-light .parse-tree-toolbar:before {
+  background: transparent;
+  border: none;
+}
+
+/*
+   This eliminates the negative left margin used to artifically bump punctuation such that it
+   doesn't appear to have space before it. This isn't worth the side-effect in SRL, as here
+   we have spans that contain things other than punctuation that are ignored. This can in the
+   long run potentially be fixed in hierplane...but that might not happen.
+*/
+.hierplane__visualization .hierplane #passage p .passage__readonly .span-slice__ignored {
+  margin-left: 0 !important;
+}
+
+/*
+  Adjust the "passage" display so that it accomodates the entire sentence, if possible. This
+  prevents the user from having to scroll when the sentence is long.
+*/
+.hierplane__visualization .hierplane--theme-light #passage {
+  height: auto !important;
+  flex: 1;
+}
+
+/*
+  Resolve wrapping issues.
+ */
+.hierplane__visualization .hierplane #passage p span {
+  display: inline;
+}
