+++ demo/src/css/icons.css
@@ -1,0 +1,28 @@
+#icons {
+  top: 0;
+  left: 0;
+  position: absolute;
+  height: 0;
+  width: 0;
+  visibility: hidden;
+  overflow: hidden;
+  z-index: -999;
+}
+
+svg {
+  fill: #000;
+}
+
+.icon__working__gradient__stop {
+  stop-color: #40affd;
+}
+
+.icon__working__gradient__stop--stop4,
+.icon__working__gradient__stop--stop5,
+.icon__working__gradient__stop--stop6 {
+  stop-opacity: 0;
+}
+
+.icon__working__path {
+  fill: url(#icon__working__gradient);
+}
