+++ demo/src/css/model.css
@@ -1,0 +1,53 @@
+.model__output {
+  background: #fff;
+}
+
+.model__output.model__output--empty {
+  background: 0 0;
+}
+
+.model__content {
+  padding: 1.875em 2.5em;
+  margin: auto;
+  -webkit-transition: padding .2s ease;
+  transition: padding .2s ease;
+}
+
+.model__content h2 {
+  margin: 0;
+  padding: 0;
+  font-size: 1em;
+}
+
+.model__content h2 span {
+  font-size: 2em;
+  color: rgba(35, 35, 35, .75);
+}
+
+.model__content h2 .tooltip,
+.model__content h2 span {
+  vertical-align: top;
+}
+
+.model__content h2 span+.tooltip {
+  margin-left: .4375em;
+}
+
+.model__content>h2:first-child {
+  margin: -.25em 0 0 -.03125em;
+}
+
+.model__content__summary {
+  font-size: 1em;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  padding: 1.25em;
+  background: #f6f8fa;
+}
+
+@media screen and (min-height:800px) {
+  .model__content {
+    padding-top: 4.6vh;
+    padding-bottom: 4.6vh;
+  }
+}
