+++ demo/src/css/visualization-types.css
@@ -1,0 +1,27 @@
+.visualization-types {
+  background: #f9fafc;
+  display: flex;
+  list-style-type: none;
+  margin: 0;
+  padding: 0 1.25em;
+  border-bottom: 1px solid #e5eaf0;
+}
+
+.visualization-types li {
+  margin: 0 1em 0 0;
+  padding: 1em 1.25em 0.5em;
+  border-bottom: 6px solid transparent;
+}
+
+.visualization-types li a {
+  color: rgba(28, 47, 58, .5);
+}
+
+
+.visualization-types .visualization-types__active-type {
+  border-bottom-color: #63a7d4;
+}
+
+.visualization-types .visualization-types__active-type a {
+  color: #1c2f3a;
+}
